# packages/create-h4ckath0n/templates/fullstack/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Standard error envelope for auth routes.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              }
            ],
            "description": "Error detail message or structured error payload.",
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PasskeyAddFinishRequest": {
        "properties": {
          "credential": {
            "additionalProperties": true,
            "description": "Browser PublicKeyCredential response as JSON.",
            "title": "Credential",
            "type": "object"
          },
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label for the device.",
            "title": "Device Label"
          },
          "device_public_key_jwk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional device public key in JWK format to bind a device identity.",
            "title": "Device Public Key Jwk"
          },
          "flow_id": {
            "description": "Flow ID returned by add/start.",
            "title": "Flow Id",
            "type": "string"
          }
        },
        "required": [
          "flow_id",
          "credential"
        ],
        "title": "PasskeyAddFinishRequest",
        "type": "object"
      },
      "PasskeyAddStartResponse": {
        "properties": {
          "flow_id": {
            "description": "Server-generated flow ID for add passkey.",
            "title": "Flow Id",
            "type": "string"
          },
          "options": {
            "additionalProperties": true,
            "description": "PublicKeyCredentialCreationOptions payload as a JSON-safe dict.",
            "title": "Options",
            "type": "object"
          }
        },
        "required": [
          "flow_id",
          "options"
        ],
        "title": "PasskeyAddStartResponse",
        "type": "object"
      },
      "PasskeyFinishResponse": {
        "properties": {
          "device_id": {
            "description": "Device ID that starts with the d prefix, empty when no device key is bound.",
            "title": "Device Id",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional display name for the user.",
            "title": "Display Name"
          },
          "role": {
            "description": "Server-side role for the user.",
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "description": "User ID that starts with the u prefix.",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "device_id",
          "role"
        ],
        "title": "PasskeyFinishResponse",
        "type": "object"
      },
...
      "PasskeyListResponse": {
        "properties": {
          "passkeys": {
            "description": "Passkeys for the current user.",
            "items": {
              "$ref": "#/components/schemas/PasskeyInfo"
            },
            "title": "Passkeys",
            "type": "array"
          }
        },
        "required": [
          "passkeys"
        ],
        "title": "PasskeyListResponse",
        "type": "object"
      },
      "PasskeyLoginFinishRequest": {
        "properties": {
          "credential": {
            "additionalProperties": true,
            "description": "Browser PublicKeyCredential response as JSON.",
            "title": "Credential",
            "type": "object"
          },
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label for the device.",
            "title": "Device Label"
          },
          "device_public_key_jwk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional device public key in JWK format to bind a device identity.",
            "title": "Device Public Key Jwk"
          },
          "flow_id": {
            "description": "Flow ID returned by login/start.",
            "title": "Flow Id",
            "type": "string"
          }
        },
        "required": [
          "flow_id",
          "credential"
        ],
        "title": "PasskeyLoginFinishRequest",
        "type": "object"
      },
...
      "PasskeyRegisterFinishRequest": {
        "properties": {
          "credential": {
            "additionalProperties": true,
            "description": "Browser PublicKeyCredential response as JSON.",
            "title": "Credential",
            "type": "object"
          },
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label for the device.",
            "title": "Device Label"
          },
          "device_public_key_jwk": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional device public key in JWK format to bind a device identity.",
            "title": "Device Public Key Jwk"
          },
          "flow_id": {
            "description": "Flow ID returned by register/start.",
            "title": "Flow Id",
            "type": "string"
          }
        },
        "required": [
          "flow_id",
          "credential"
        ],
        "title": "PasskeyRegisterFinishRequest",
        "type": "object"
      },
      "PasskeyRegisterStartRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional display name for the new user account.",
            "title": "Display Name"
          }
        },
        "title": "PasskeyRegisterStartRequest",
        "type": "object"
      },
      "PasskeyRegisterStartResponse": {
        "properties": {
          "flow_id": {
            "description": "Server-generated flow ID for registration.",
            "title": "Flow Id",
            "type": "string"
          },
          "options": {
            "additionalProperties": true,
            "description": "PublicKeyCredentialCreationOptions payload as a JSON-safe dict.",
            "title": "Options",
            "type": "object"
          }
        },
        "required": [
          "flow_id",
          "options"
        ],
        "title": "PasskeyRegisterStartResponse",
        "type": "object"
      },
      "PasskeyRenameRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New passkey name. Null or empty to clear.",
            "title": "Name"
          }
        },
        "title": "PasskeyRenameRequest",
        "type": "object"
      },
      "PasskeyRenameResponse": {
        "properties": {
          "id": {
            "description": "Internal passkey ID.",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated passkey name.",
            "title": "Name"
          }
        },
        "required": [
          "id"
        ],
        "title": "PasskeyRenameResponse",
        "type": "object"
      },
      "PasskeyRevokeResponse": {
        "properties": {
          "message": {
            "description": "Status message for the revocation action.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "PasskeyRevokeResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "DeviceJWT": {
        "description": "Device-signed ES256 JWT minted by the client. The JWT header must include kid set to the device id, and the aud claim must be h4ckath0n:http.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Hackathon app powered by h4ckath0n",
    "title": "h4ckath0n",
    "version": "0.1.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/passkey/add/finish": {
      "post": {
        "description": "Verify the WebAuthn attestation and attach the new passkey to the user.",
        "operationId": "add_finish_auth_passkey_add_finish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyAddFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyFinishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or expired flow, or WebAuthn verification error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Finish adding a passkey",
        "tags": [
          "passkey"
        ]
      }
    },
    "/auth/passkey/add/start": {
      "post": {
        "description": "Begin adding a new passkey for the authenticated user and return registration options.",
        "operationId": "add_start_auth_passkey_add_start_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyAddStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid token."
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Start adding a passkey",
        "tags": [
          "passkey"
        ]
      }
    },
    "/auth/passkey/login/finish": {
      "post": {
        "description": "Finish passkey login by verifying the WebAuthn assertion for the flow and binding an optional device key.",
        "operationId": "login_finish_auth_passkey_login_finish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyLoginFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyFinishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid credentials, revoked passkey, or expired flow."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finish passkey login",
        "tags": [
          "passkey"
        ]
      }
    },
...
    "/auth/passkey/register/finish": {
      "post": {
        "description": "Finish passkey registration by verifying the WebAuthn attestation for the flow and binding an optional device key.",
        "operationId": "register_finish_auth_passkey_register_finish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegisterFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyFinishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or expired flow, or invalid WebAuthn payload."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finish passkey registration",
        "tags": [
          "passkey"
        ]
      }
    },
    "/auth/passkey/register/start": {
      "post": {
        "description": "Begin a passkey registration ceremony. Creates a new user and a single-use challenge flow, then returns WebAuthn registration options.",
        "operationId": "register_start_auth_passkey_register_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRegisterStartRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyRegisterStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or WebAuthn configuration error."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start passkey registration",
        "tags": [
          "passkey"
        ]
      }
    },
    "/auth/passkeys": {
      "get": {
        "description": "List all passkeys, including revoked entries, for the current user.",
        "operationId": "passkeys_list_auth_passkeys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid token."
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "List passkeys",
        "tags": [
          "passkey"
        ]
      }
    },
    "/auth/passkeys/{key_id}": {
      "patch": {
        "description": "Update the user-facing name of a passkey.",
        "operationId": "passkey_rename_auth_passkeys__key_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyRenameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Passkey not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cannot rename a revoked passkey."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error."
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Rename a passkey",
        "tags": [
          "passkey"
        ]
      }
    },
    "/auth/passkeys/{key_id}/revoke": {
      "post": {
        "description": "Revoke a passkey by its internal key ID. The last active passkey cannot be revoked and returns the LAST_PASSKEY error.",
        "operationId": "passkey_revoke_auth_passkeys__key_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyRevokeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Passkey not found."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "LAST_PASSKEY",
                    "message": "Cannot revoke the last active passkey."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cannot revoke the last active passkey."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DeviceJWT": []
          }
        ],
        "summary": "Revoke a passkey",
        "tags": [
          "passkey"
        ]
      }
    },
...
  }
}